+++ Functions/weierstrass.h
@@ -11,10 +11,10 @@
 	#define MIN_SEARCH_RANGE -0.5
 	#define MAX_SEARCH_RANGE 0.5
 
-	float formulae(float values[]);
-		float firstCalculation(float values[]);
-		float secondCalculation();
-		float calcolateSummation(float num);
+	double formulae(double values[]);
+		double firstCalculation(double values[]);
+		double secondCalculation();
+		double calcolateSummation(double num);
 
 #endif
 #endif
